+++ game/cards/dm05/gel_fish.go
@@ -59,3 +59,67 @@ func SplitHeadHydroturtleQ(c *match.Card) {
 	})
 
 }
+
+// LurkingEel ...
+func LurkingEel(c *match.Card) {
+
+	c.Name = "Lurking Eel"
+	c.Power = 4000
+	c.Civ = civ.Water
+	c.Family = family.GelFish
+	c.ManaCost = 6
+	c.ManaRequirement = []string{civ.Water}
+
+	c.Use(func(card *match.Card, ctx *match.Context) {
+
+		if event, ok := ctx.Event.(*match.AttackPlayer); ok {
+
+			blockers := make([]*match.Card, 0)
+
+			attacker, err := ctx.Match.Opponent(c.Player).GetCard(event.CardID, match.BATTLEZONE)
+
+			if err != nil {
+				return
+			}
+
+			if attacker.Civ == civ.Fire || attacker.Civ == civ.Nature {
+				return
+			}
+
+			for _, b := range event.Blockers {
+				if b.ID != card.ID {
+					blockers = append(blockers, b)
+				}
+			}
+
+			event.Blockers = blockers
+
+		}
+
+		if event, ok := ctx.Event.(*match.AttackCreature); ok {
+
+			blockers := make([]*match.Card, 0)
+
+			attacker, err := ctx.Match.Opponent(c.Player).GetCard(event.CardID, match.BATTLEZONE)
+
+			if err != nil {
+				return
+			}
+
+			if attacker.Civ == civ.Fire || attacker.Civ == civ.Nature {
+				return
+			}
+
+			for _, b := range event.Blockers {
+				if b.ID != card.ID {
+					blockers = append(blockers, b)
+				}
+			}
+
+			event.Blockers = blockers
+
+		}
+
+	}, fx.Creature, fx.Blocker)
+
+}